+++ lib/app/auth/instance/join/join_auth_instance_bloc.dart
@@ -1,4 +1,5 @@
 import 'package:easy_dispose/easy_dispose.dart';
+import 'package:fedi/app/server_list/server_list_auto_complete_bloc.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 
@@ -8,7 +9,11 @@ abstract class IJoinAuthInstanceBloc implements IDisposable {
 
   bool get isFromScratch;
 
+  IServerListAutoCompleteBloc get serverListAutoCompleteBloc;
+
   TextEditingController get hostTextController;
+
+  FocusNode get hostFocusNode;
 
   Uri extractCurrentUri();
 }