+++ lib/app/auth/instance/join/join_auth_instance_bloc_impl.dart
@@ -2,7 +2,10 @@ import 'package:easy_dispose/easy_dispose.dart';
 import 'package:easy_dispose_flutter/easy_dispose_flutter.dart';
 import 'package:easy_dispose_provider/easy_dispose_provider.dart';
 import 'package:fedi/app/auth/instance/join/join_auth_instance_bloc.dart';
+import 'package:fedi/app/auth/instance/join/join_auth_instance_bloc_proxy_provider.dart';
 import 'package:fedi/app/config/config_service.dart';
+import 'package:fedi/app/server_list/server_list_auto_complete_bloc.dart';
+import 'package:fedi/app/server_list/server_list_auto_complete_bloc_impl.dart';
 import 'package:flutter/cupertino.dart';
 
 class JoinAuthInstanceBloc extends DisposableOwner
@@ -11,6 +14,12 @@ class JoinAuthInstanceBloc extends DisposableOwner
   final bool isFromScratch;
   @override
   final TextEditingController hostTextController = TextEditingController();
+  @override
+  final FocusNode hostFocusNode = FocusNode();
+
+  @override
+  final IServerListAutoCompleteBloc serverListAutoCompleteBloc =
+      ServerListAutoCompleteBloc();
 
   final IConfigService configService;
 
@@ -19,11 +28,16 @@ class JoinAuthInstanceBloc extends DisposableOwner
     required this.configService,
   }) {
     hostTextController.disposeWith(this);
+    hostFocusNode.disposeWith(this);
+
+    // don't need to await we start init but don't need wait to finish
+    serverListAutoCompleteBloc.performAsyncInit();
+    serverListAutoCompleteBloc.disposeWith(this);
   }
 
   @override
   Uri extractCurrentUri() {
-    var uriText = hostTextController.text;
+    var uriText = hostTextController.text.trim();
 
     if (!uriText.isNotEmpty) {
       uriText = configService.appDefaultInstanceUrl;
@@ -64,7 +78,9 @@ class JoinAuthInstanceBloc extends DisposableOwner
         context,
         isFromScratch: isFromScratch,
       ),
-      child: child,
+      child: JoinAuthInstanceBlocProxyProvider(
+        child: child,
+      ),
     );
   }
 }